+++ src/main/java/org/embulk/util/config/modules/ColumnJacksonModule.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.embulk.spi;
+package org.embulk.util.config.modules;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.core.JsonParser;
@@ -32,6 +32,7 @@ import java.io.IOException;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
+import org.embulk.spi.Column;
 import org.embulk.spi.type.Type;
 import org.embulk.spi.type.Types;
 import org.slf4j.Logger;
