+++ src/main/java/org/embulk/util/config/modules/TypeJacksonModule.java
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 
-package org.embulk.spi.type;
+package org.embulk.util.config.modules;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import java.io.IOException;
+import org.embulk.spi.type.Type;
 
 public final class TypeJacksonModule extends SimpleModule {
     public TypeJacksonModule() {
